# functions/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "FireStock API",
    "description": "The official FireStock API docs"
  },
  "host": "europe-west1-firestock-cloud.cloudfunctions.net/api",
  "basePath": "/v1",
  "tags": [ 
    { 
      "name": "Authentification", 
      "description": "Création de compte, connexion, récupération ..."
    },
    {
      "name": "Bucket", 
      "description": "Lister, uploader, télécharger et supprimer des fichiers"
    }
  ],
  "schemes": ["https"],
  "paths": { 
    "/auth/register": { 
      "post": { 
        "tags": ["Authentification"], 
        "summary": "Créer un nouveau compte",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Register" },
            "description": "Paramètres de création d'un nouvel utilisateur"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Utilisateur créé avec succès !"
          },
          "400": { 
            "description": "Erreur lors de la création de l'utilisateur !"
          } 
        } 
      } 
    },
    "/auth/login": { 
      "post": { 
        "tags": ["Authentification"], 
        "summary": "Connecter un utilisateur",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Login" },
            "description": "Paramètres de connexion d'un utilisateur"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Renvoi le token de session"
          },
          "400": { 
            "description": "Erreur lors de la connexion de l'utilisateur !"
          } 
        } 
      } 
    },
    "/auth/recovery": { 
      "post": { 
        "tags": ["Authentification"], 
        "summary": "Réinitialiser le mot de passe",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Recovery" },
            "description": "Paramètres de récupération d'un utilisateur"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Email de récupération envoyé avec succès !"
          },
          "400": { 
            "description": "Erreur lors de la récupération du compte !"
          } 
        } 
      } 
    },
    "/auth/refresh": { 
      "post": { 
        "tags": ["Authentification"], 
        "summary": "Rafraîchir le token de session",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Refresh" },
            "description": "Paramètres de rafraîchissement d'un token"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Nouveau token généré avec succès !"
          },
          "400": { 
            "description": "Erreur lors du rafraîchissement du token !"
          } 
        } 
      } 
    },
    "/auth/logout": { 
      "get": { 
        "tags": ["Authentification"], 
        "summary": "Déconnecter l'utilisateur",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": { 
          "200": { 
            "description": "Utilisateur déconnecté avec succès !"
          },
          "400": { 
            "description": "Erreur lors de l'invalidation des tokens !"
          } 
        } 
      } 
    },
    "/bucket/data": { 
      "get": { 
        "tags": ["Bucket"], 
        "summary": "Récupérer les données de l'utilisateur",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Chemin du dossier à récupérer"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Données récupérées avec succès !"
          },
          "400": { 
            "description": "Erreur lors de la récupération des données !"
          } 
        } 
      } 
    },
    "/bucket/download": { 
      "get": { 
        "tags": ["Bucket"], 
        "summary": "Télécharger une donnée",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Chemin du fichier à télécharger"
          }
        ],
        "responses": { 
          "200": { 
            "description": "Lien de téléchargement récupéré avec succès !"
          },
          "400": { 
            "description": "Erreur lors de la récupération du lien !"
          } 
        } 
      } 
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Register": {
      "required": ["email", "password", "pseudo", "firstname", "lastname", "phone"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "pseudo": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        }
      }
    },
    "Login": {
      "required": ["pseudo", "password"],
      "properties": {
        "pseudo": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Recovery": {
      "required": ["pseudo"],
      "properties": {
        "pseudo": {
          "type": "string"
        }
      }
    },
    "Refresh": {
      "required": ["refreshToken"],
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    }
  }
}